+++ tests/testthat/test-tt_save_dataset.R
@@ -1,0 +1,11 @@
+test_that("tt_save_dataset creates the expected files", {
+  skip_on_ci()
+  proj_dir <- withr::local_tempdir()
+  full_submission_path <- fs::path(proj_dir, "tt_submission")
+  usethis::local_project(proj_dir, force = TRUE, quiet = TRUE)
+  expect_true(tt_save_dataset(mtcars, open = FALSE))
+  file_path <- fs::path(full_submission_path, "mtcars.csv")
+  expect_snapshot_file(file_path)
+  file_path <- fs::path(full_submission_path, "mtcars.md")
+  expect_snapshot_file(file_path)
+})